* commit '58e7a0044bcc2fbb99061dd9be5108aa94d07223':
  new file:   src/main/java/com/oracle/athena/webserver/buffermgr/BufferAssociation.java 	modified:   src/main/java/com/oracle/athena/webserver/http/CasperHttpInfo.java 	modified:   src/main/java/com/oracle/athena/webserver/memory/FixedSizeBufferPool.java 	modified:   src/main/java/com/oracle/athena/webserver/memory/MemoryManager.java 	modified:   src/main/java/com/oracle/athena/webserver/niosockets/NioSocket.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/BufferReadMetering.java 	new file:   src/main/java/com/oracle/athena/webserver/operations/BuildHeaderToStorageServer.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/ConnectComplete.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/EncryptBuffer.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/OperationTypeEnum.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/ParseHttpRequest.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/ReadBuffer.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/SetupChunkWrite.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/SetupV2Put.java 	new file:   src/main/java/com/oracle/athena/webserver/operations/WriteHeaderToStorageServer.java 	modified:   src/main/java/com/oracle/athena/webserver/operations/WriteToStorageServer.java 	modified:   src/main/java/com/oracle/athena/webserver/requestcontext/RequestContext.java 	modified:   src/test/java/com/oracle/athena/webserver/client/ClientConnectComplete.java 	modified:   src/test/java/com/oracle/athena/webserver/client/ClientHttpRequestWrite.java 	modified:   src/test/java/com/oracle/athena/webserver/client/ClientObjectWrite.java 	modified:   src/test/java/com/oracle/athena/webserver/client/ClientResponseHandler.java 	modified:   src/test/java/com/oracle/athena/webserver/client/ClientWrite.java 	modified:   src/test/java/com/oracle/athena/webserver/client/SetupClientConnection.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/ClientTest.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/ClientTest_BadMd5.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/ClientTest_CheckMd5.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/ClientTest_InvalidMd5Header.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/ClientTest_MissingObjectName.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/TestEncryptBuffer.java 	modified:   src/test/java/com/oracle/athena/webserver/manual/TestMain.java

+++ src/main/java/com/oracle/athena/webserver/http/CasperHttpInfo.java
@@ -95,7 +95,7 @@ public class CasperHttpInfo {
 
     private String UserAgent;
 
-    private long contentLength;
+    private int contentLength;
 
     /*
      **
@@ -462,7 +462,7 @@ public class CasperHttpInfo {
         if (result != -1) {
             try {
                 contentLengthReceived = true;
-                contentLength = Long.parseLong(field.getValue());
+                contentLength = int.parseLong(field.getValue());
 
                 /*
                  ** TODO: Are there specific limits for the Content-Length that need to be validated
